extract-info-from-images/requirements.txt
openai>=1.3.0
azure-identity>=1.15.0
azure-ai-projects==1.0.0b12
azure-ai-agents==1.1.0b4
azure-core>=1.29.0
pillow>=10.0.0
click>=8.0.0
python-dotenv>=1.0.0
